requirements.txt
autopep8==1.7.0
bidict==0.22.0
blinker==1.5
cachelib==0.9.0
click==8.1.3
dnspython==2.2.1
Flask==2.1.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SocketIO==4.3.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycodestyle==2.9.1
python-decouple==3.6
python-dotenv==0.20.0
python-engineio==3.13.2
python-socketio==4.6.0
six==1.16.0
SQLAlchemy==1.4.40
toml==0.10.2
typing_extensions==4.3.0
Werkzeug==2.0.1
WTForms==3.0.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
